fixed scattering normalizations
This also brings the produced phase curves in line with the derivations in the notes - the notes have correspondingly been updated. This was done in parallel with improvements to the normalization of optool phase functions - wavelength_testing is currently comparing a fully optool-based ring to diffuse scattering with an optool opacity.